update new monitor snapshot
